split test_macros
